+++ CHANGE_LOG.md
@@ -6,6 +6,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed font family issue with pre-chat survey pane
+
 ## [1.1.0] - 2023-6-8
 
 ### Added